backend/requirements.txt
absl-py==2.1.0
APScheduler==3.10.4
asgiref==3.8.1
astunparse==1.6.3
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
djangorestframework==3.15.2
firebase-admin==6.5.0
flatbuffers==24.3.25
gast==0.6.0
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
h5py==3.11.0
httplib2==0.22.0
idna==3.7
joblib==1.4.2
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
msgpack==1.0.8
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.7.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.14.0
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.1
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
<<<<<<< HEAD
wrapt==1.16.0
=======
wrapt==1.16.0
>>>>>>> fb3ee44414eba0c16c75ca057fceca47299d2f24
